requirements.txt
Babel==2.8.0
certifi==2020.4.5.2
chardet==3.0.4
Django==2.2.13
django-extra-views==0.11.0
django-haystack==2.8.1
django-localflavor==3.0.1
django-oscar==2.0.4
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
factory-boy==2.12.0
Faker==4.1.0
idna==2.9
phonenumbers==8.12.5
Pillow==7.1.2
psycopg2==2.8.5
purl==1.5
pycountry==19.8.18
python-dateutil==2.8.1
python-stdnum==1.13
pytz==2020.1
requests==2.23.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
